requirements.txt
psutil
redis
psycopg2-binary
pandas
numpy
scikit-learn
matplotlib
fastapi
uvicorn
joblib